Split market.rs (617 lines) into market/ subdirectory
Split crates/simulation/src/market.rs into focused modules under market/:
- events.rs (114 lines): MarketEvent enum, ActiveMarketEvent
- types.rs (125 lines): PriceEntry, MarketPrices resource
- pricing.rs (250 lines): pseudo-random helpers, update_market_prices system
- plugin.rs (17 lines): MarketPlugin
- tests.rs (156 lines): all unit tests
- mod.rs (10 lines): thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
