+++ crates/simulation/src/market/mod.rs
@@ -1,0 +1,10 @@
+mod events;
+mod plugin;
+mod pricing;
+mod tests;
+mod types;
+
+pub use events::{ActiveMarketEvent, MarketEvent};
+pub use plugin::MarketPlugin;
+pub use pricing::update_market_prices;
+pub use types::{MarketPrices, PriceEntry};
